backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==10.3.0
pypng==0.20220715.0
qrcode==7.4.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
user-agents==2.2.0
geoip2